backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
yt_dlp
python-dotenv
